python: simplify
